Moved lsp-augment.el to the root of the repo
Because when I have this config:

(use-package lsp-augment
  :after (lsp-mode markdown-mode)
  :vc (:fetcher github :repo rolandd/augment.vim)
  ...

Emacs byte compiles the lsp-augment.el file, and creates other files
that are left in the root of the repo, but don't point to the .elc
file.

With this crude change, I want to see if this will just check out the
file and automatically add it to load-path.