Allow chat messaging from the chat too
Previoulsy, only chat messages could be sent from a souce code file
under a LSP session.

This change adds buffer local variables so that it knows what LSP
session its linked back to.

Why is this relevant?

Because this mode works via LSP and since the chat buffers aren't
under a LSP session, trying to send the chat messages wouldn't work.

By assingning the "parent" LSP session to a buffer local variable,
that information is retained and available for the buffer. That way it
knows what LSP session to send Augment Code LSP specific messages to.