comint shell with prompt added, rough
I've got the proof of concept working. It's a bit rough & will need
polish.

But what is there right now?

The original, send messages via minibuffer & append response to a
un-writable buffer is still in code, well, just for compatibility.
I doubt that many people are using it and loving it as is.

Figured that having a comint based shell with prompt would be much
"more natural" way to interact with the chat.

So that's what I implemented.

I kept the old *Augment Chat History* buffer name for the append to
buffer chat. But used a different buffer name for the comint buffer.

When this is cleaned up, it'll be possible to have multiple chats,
concurrently. Useful when you get in those "I tried to solve one
problem, now I've got two more problems" situations. So that you can
leave one train of thought in the original chat and then you can start
a separate chat where you can go off on a side quest, to solve those
pesky bugs.